# Conflicts:
#	retail-auth/src/main/java/com/retail/auth/controller/AuthController.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/BargainEntity.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/BargainRecordEntity.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/MateEntity.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/SeckillEntity.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/TeamEntity.java
#	retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/TeamRecordEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/ColonelApplyEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/ColonelEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/GradeEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/RouteEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/src/main/java/com.retail.colonel/domain/WarehouseEntity.java
#	retail-modules/retail-colonel/retail-colonel-common/target/retail-colonel-common-1.0-SNAPSHOT.jar
#	retail-modules/retail-colonel/retail-colonel-remote/target/retail-colonel-remote-1.0-SNAPSHOT.jar
#	retail-modules/retail-colonel/retail-colonel-server/target/retail-colonel-server-1.0-SNAPSHOT.jar
#	retail-modules/retail-order/retail-order-common/src/main/java/com.retail.order/domain/ExpressageEntity.java
#	retail-modules/retail-order/retail-order-common/src/main/java/com.retail.order/domain/OrderEntity.java
#	retail-modules/retail-order/retail-order-common/src/main/java/com.retail.order/domain/ReatilOrderInfoEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/BrandEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/ClassifiedEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/FirstCycleImageEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/RetaiUserCouponCenterEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/ShopEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/SkuEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/SpecificationEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/SpuEntity.java
#	retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/StoreCouponEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/IntegrationHistoryEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/RoleEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/UserAddressEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/UserCollectShopEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/UserCollectSkuEntity.java
#	retail-modules/retail-user/retail-user-common/src/main/java/com/retail/user/domain/UserRoleEntity.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/controller/RoleController.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/controller/UserAddressController.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/service/RoleService.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/service/UserAddressService.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/service/impl/RoleServiceImpl.java
#	retail-modules/retail-user/retail-user-server/src/main/java/com/retail/user/service/impl/UserAddressServiceImpl.java

+++ retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/BargainEntity.java
@@ -1,5 +1,4 @@
 package com.retail.bargain.domain;
-
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
@@ -73,6 +72,6 @@ public class BargainEntity implements Serializable {
 	/**
 	 * 状态
 	 */
-	private Integer status;
+	private Long status;
 
 }
