+++ retail-modules/retail-shop/retail-shop-common/src/main/java/com/retail/shop/domain/ClassifiedEntity.java
@@ -1,5 +1,4 @@
 package com.retail.shop.domain;
-
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
@@ -33,7 +32,7 @@ public class ClassifiedEntity implements Serializable {
 	/**
 	 * 父级id
 	 */
-	private Long pid;
+	private Integer pid;
 	/**
 	 * 分类名称
 	 */
